Relocate properties to DBSpec for further sharing
